requirements.txt
pymupdf
pymupdf4llm
pymupdf-layout
rapidocr-onnxruntime
pytesseract
Pillow
docling
pandas
